--- src/components/image-modal/has-breed-content/index.ts
@@ -1,3 +1,0 @@
-import HasBreedContent from "./has-breed-content";
-
-export default HasBreedContent;
