--- src/components/image-modal/index.ts
@@ -1,3 +1,0 @@
-import ImageModal from "./image-modal";
-
-export default ImageModal;
